general updating

+++ apps/schemas/nilai_akhir_schema.py
@@ -3,7 +3,7 @@ from pydantic import BaseModel
 class NilaiAkhirSchema(BaseModel):
     usersid: str
     kd_matkul: str
-    nilai_akhir: int
+    nilai_akhir: float
 
     class Config:
         orm_mode = True
